requirements.txt
flask
numpy
opencv-python
pillow
rasterio
tensorflow